Remove unnecessary main component
